feat(graphql): [#55963] added class_delete mutator
This change introduces the `class_delete` mutator used to do bitemporal
deletion, not to be confused with `class_terminate` which does temporal
'deletion', aka terminates the validity for a class by a given date.

As this mutator works on the bitemporal time-axis, it should only be
used by clients that fully understand the underlying bitemporal model,
including how a bitemporal delete affects the registration history.

In order to do bitemporal deletion using this mutator, simply call it
with a UUID like below:
```graphql
mutation DeleteClass {
  class_delete(uuid: "4e337d8e-1fd2-4449-8110-e0c8a22958ed") {
    uuid
  }
}
```
After this call the class will no longer show up in any temporal
listing. Do note that doing bitemporal deletions on entities that are
referenced will leave dangling references to the class potentially
breaking invariants in the data.

It is currently the callers responsibility to ensure that references are
dealt with before doing bitemporal deletions.