feat(graphql): [#55975] added class_update mutator
This change introduces the `class_update` mutator used to update
classes. The mutator takes two arguments, the UUID of the class to
update and a ClassUpdateInput, which is currently identical to the
ClassCreateInput.

The update endpoint cannot be used to create new classes, it can only be
used to update or reactivate existing classes.

In order to update a class using this mutator, simply call it like below:
```graphql
mutation UpdateClass {
  class_update(
    uuid: "e75f74f5-cbc4-4661-b9f4-e6a9e05abb2d",
    input: {
      name: "Postal Address",
      user_key: "AdressePostEmployee",
      facet_uuid: "5b3a55b1-958c-416e-9054-606b2c9e4fcd"
    }
  ) {
    uuid
  }
}
```
After this call the class will now have the specified data.