test(oio_rest): [#55973] removed unnecessary and flaky test
This change removes the `test_build_registration_dokument` testcase,
as it is a flaky and unnecessary test. The flakiness occured as a result
of the partial removal of the dokument data type from LoRa in version
3.0.0, and has been sporadically failing ever since.

As the dokument data type has mostly been removed, removing this test
should not negatively affect anyone.