chore: [#55975] change gunicorn / uvicorn initialization to application factory
This commit changes our gunicorn / uvicorn initialization to use the
application factory pattern via the `create_app` function, instead of
having a `main.py` file, with the sole purpose of importing and calling
the `create_app` function.

The application factory pattern is preferable as it forces users of the
application to consider the lifecycle of the application they create and
as it decentivize the anti-pattern of using global variables.