feat(graphql): [#55760] remove cursor from v4 and offset from v5
This commit removes the `cursor` argument from GraphQL version 4 and the
`offset` parameter from GraphQL version 5, as GraphQL version 4 did
indexed-based pagination while GraphQL version 5 does cursor-based
pagination and as we do not want to mix the two styles.

An `offset` to `cursor` translation layer has been implemented in the
GraphQL version 4 implementation, such that the latest version / GraphQL
version 5 only has to support cursor-based pagination, this is done
using insider information on the format of the cursors.