feat(graphql): [#55976] added facet_update mutator
This change introduces the `facet_update` mutator used to update
facets. The mutator takes two arguments, the UUID of the facet to
update and a FacetUpdateInput, which is currently identical to the
FacetCreateInput.

The update endpoint cannot be used to create new facets, it can only be
used to update or reactivate existing facets.

In order to update a facet using this mutator, simply call it like below:
```graphql
mutation UpdateFacet {
  facet_update(
    uuid: "e75f74f5-cbc4-4661-b9f4-e6a9e05abb2d",
    input: {
      user_key: "EmployeeAddressType"
    }
  ) {
    uuid
  }
}
```
After this call the facet will now have the specified data.