feat(graphql): [#56049] added mutator roles
This commit adds `create`, `update`, `terminate`, `delete` and `refresh`
role checks to the appropriate mutators, replacing the previous `admin`
role check that was protecting all mutators before.

As such the `admin` role is no longer special with regards to GraphQL
mutators, but rather it is just a composite role, which happens to
contain all of the direct low-level roles.

This change is intended to be entirely non breaking, as the admin role
should already contain all the direct low-level roles as its composites.