fix: [#56201] defer owner/user checking
Before, the Keycloak token would immediately be rejected if it had the
'owner' role, but didn't also supply a user uuid. This caused issues
with the new Keycloak realm builder[1], as the 'admin' role was changed
from being a role in OS2mo to a composite role directly in keycloak.
Whereas the previous implementation supplied MO with a token containing
the 'admin' role, the new realm builder sends a token containing all
create_, delete_ etc. roles, _including_ the 'owner' role. We allow
tokens to have the 'owner' role, without also supplying a user UUID,
until it is needed to verify ownership.

[1] https://git.magenta.dk/rammearkitektur/keycloak-realm-builder/-/merge_requests/66