test!: [#55760] remove the LoRa testing API
This commit removes the LoRa testing API, that was previously available
on `/lora/testing` if the `testing_api` environment variable was set to
`True`.

BREAKING CHANGE:

As this commit removes a public facing interface it is a breaking change.

It is however our belief that noone uses this API and thus no migration
path is provided.