fix: [#55042] allow full access to admins
Backwards-compatible fix for !1594. Before OS2mo v21.12.0, the 'admin'
role was checked directly in MO. After v21.12.0, the role was made into
a composite role in keycloak, and as such handling of the 'admin' was
removed from MO in favour of checking the specific roles individually.
This is only a backwards-compatible change in the context of the whole
stack, including the keycloak realm builder, which maintains these
roles. Aarhus, however, manually manages their (our!) keycloak
configuration, and as such the change _is_ breaking in their
environment.