fix: [#56654] replace file download authorization cookie
OS2mo 21.12.0 unintentionally broke the file downloads for some users.

File downloads were broken as a consequence of two seemingly unrelated
facts, namely:
1. The 4KB limit on cookies (RFC2109 (6.3), RFC2965 (5.3), and RFC6265)
2. The introduction of new roles in the role-based access control system.

Before this commit the file download authorization cookie was set to the
JWT token as provided by Keycloak. This allowed the backend to verify the
token on download using the Keycloak public signing key.

However as new roles were introduced the size of the JWT tokens grew,
eventually surprissing the 4KB limit on cookies, thus the web-browser
would reject the cookie thereby breaking file downloads.

This fix modifies the value returned in the file download authorization
cookie to be a random UUID instead of the JWT token, thus ensuring that
the output size is fixed. The cookie UUID will be valid for 10 minutes.

This change also adds a new database table to keep track of the file
tokens.