feat(graphql): [#50473] introduce owner reading to GraphQL
This commit introduces reading of owner objects to our GraphQL interface.

It should be noted that owner objects are not implemented as fully yet.
This change only introduces the object itself, not all its realations.

However it is not possible to run queries alike these:
```graphql
{
  org_units {
    objects {
      current {
        owners {
          uuid
        }
      }
    }
  }
}
```
Or:
```graphql
{
  owners {
    objects {
      current {
        uuid
      }
    }
  }
}
```
And to check if a given organisation unit has owners attached or not.