feat(graphql): [#56757] allow setting nickname when creating employees
Before this commit it was not possible to set nickname when creating an
employee using the `employee_create` mutator. A workaround was however
possible as the `employee_update` mutator allowed for setting it.

Thus it was possible to create an employee without a nickname using the
`employee_create` mutator and then immediately add a nickname using the
`employee_update` mutator, however this workflow is cumbersome and
includes the risk of having the employee in OS2mo without a nickname for
a certain amount of time.

This commit does not solve the general issue of mismatches between
create and update operations, but rather solves the specific issue with
nicknames. In the future all these distinctions should be dealt with,
such that whatever can be edited after creation also can be set during
creation.