fix(graphql): [#53111] change mutator input-name from 'at' to 'input, to align with other mutators
This commit introduces a new GraphQL version, GraphQL v8.
I stumbled upon the name of an input on the `address_terminate`-mutator,
which didn't align with the other mutators and changed it from 'at' to 'input'.

To migrate to the new version, change all current GraphQL calls from
version 7 (`/graphql/v7`) to version 8 (`/graphql/v8`) and change the variable name
for the mutator `address_terminate` from `at` to `input`.