feat: [#56849] add expose_lora flag to OS2mo
Before this change **all** OS2mo instances would **always** expose
**all** of LoRas REST API on the `/lora` prefix.

This change makes this behavior configurable, such that deployments can
decide whether or not they want to expose LoRa.
This is the first step in the progress of eliminating LoRas REST API, in
line with the goal of only exposing OS2mo data through the GraphQL
interface.