feat(auth): [#56940] implement set_authenticated_user
This commit ensures that all changes written to registrations in the
database have the correct user references.

Before this change all registration user references would simply be
set to: `42c432e8-9c4a-11e6-9f62-873cf34a735f`, a "magic number" UUID
that was introduced to OS2mo/LoRa in December of 2015 and has been
used since then as the only user reference.

Thus after this commit the "magic number" UUID should appear much less
frequently, although it is still used for backwards compatibility for
occasions when the user reference cannot be extracted from the JWT token
from Keycloak.

The implementation has been written using `starlette-context` to pass
the user reference outside of the stack, to get it into LoRa's database
layer without actually modifying the entire call stack. It has been
implemented in accordance with:

* https://starlette-context.readthedocs.io/en/latest/fastapi.html

In the future when shimming has been implemented fully, and LoRa's REST
API has been eliminated a better solution can probably be made by
passing it inside the call stack, but this is the solution for now.