feat: [#xxxxx] ensure container image reproducibility
This change ensures that the OS2mo container image can built in a
reproductive way, i.e. that the build process is deterministic and that
two consecutive builds produce two binarily identical container images.

For details on why this is desirable please check:

* https://reproducible-builds.org/

The reproducible images can be made using:
```console
$ kaniko build \
    --reproducible
    --build-arg SOURCE_DATE_EPOCH=(git log -1 --pretty=%ct)
```
The above example uses `kaniko` instead of `docker` to build the image,
as there is no way to set the layer timestamp with `docker` (which
causes the hash to vary), `kaniko` has the `--reproducible` flag to work
around this issue (`buildah` has a similar `--timestamp` flag).

The reproducibility is achieved in Python by means of `SOURCE_DATE_EPOCH`
with PEP-552 being the relevant PEP standard.

Reproducibility is subject to all dependencies being versioned, and thus
on the elimination of always latest "git" sourced dependencies.