perf: [#56537] Removed ServiceType from code base
The new amqp-model only provides an objectType and uuid. So we need to
find the service type without using the routing key.

This MR removes the ServiceType object from the code base and provides a
workaround to retrieve the service type. Hence it will be easier to
transfer to the new AMQP system