feat(graphql): [#57145] align employee create + update mutators
This commit ensures that the same arguments are taken by the
`employee_create` and `employee_update` mutators.

As a side-effect both mutators now take, both:
* `cpr_number` and `cpr_no`
* `given_name` and `givenname`
Of which the latters are being deprecated and subject to removal in
the future.

Additionally both mutators also now take `name` and `nickname`, both of
which are deprecated in favor of their structured counterparts.