feat(graphql): [#57145] remove type input from the `ituser_manager` mutator
This commit introduces a new GraphQL version, GraphQL v12.

GraphQL version 12 introduces a breaking change to the input variables taken
by the `ituser_create`-mutator. Specifically it removes the `type` input
variable, which was previously an optional argument that should *always* have
its value set to: `it`. Setting it to any other value will break
invariants in the underlying code leading to undefined behavior.

The argument had the default value of `it` and as such the issue
only arises if the caller explicitly sends a different value.
The argument has however been removed entirely in version 12, as it
is leaking implementation-specific details and should never have been
exposed.

To migrate from GraphQL version 11, simply stop sending `type` with your queries.