test: [#57488] reimplement truncate_all_tables helper
This commit reimplements the `truncate_all_tables` test helper using
SQLAlchemy operations replacing the previous implementation using raw
sql statements.

The change was made to ensure interoperability with the rest of the
fixtures in the test-suite, in particular issues related to using the
`empty_db` fixture in OS2mo's test-suite.