feat(audit)!: [#57568] require auditor and registration roles
This commit makes the `read_auditlog` role required to access the
`auditlog` top-level GraphQL object and the `read_registration` role
required to access the `registrations` top-level GraphQL object.

BREAKING CHANGE:

While this is indeed a breaking change, as new permissions are now
required for the `auditlog` and `registrations`, it should be
non-breaking for most users as a corresponding change has been made to
the Keycloak realm builder version 4.4.0.