fix(audit): [#57596] fix bad context manager usage in autocomplete
This commit fixes the `/e/autocomplete/` and `/o/autocomplete`
endpoints, when `CONFDB_AUTOCOMPLETE_V2_USE_LEGACY` is disabled.
These endpoints were unintentionally broken by OS2mo 26.4.0.

The root-cause of the issue was a missing test-case for the endpoint.
A couple of test-cases have been added to avoid similar regressions in the future.