fix: [#57650] don't set empty manager responsibility/level
The issue was caused by updating exactly one of manager responsibility
or level but not the other. See the comment in the committed code for
details. This is (currently) only a problem for manager's responsibility
and level, as they are the only MO fields that share the same LoRa
field.

This issue was known and documented (in the tests) since
c28cd16899a1be280ae6132006bba05bff4d7ec8, but apparently was not
disallowed in the GraphQL schema.