feat!: [#57816] allow infinite IT-systems
If we don't, existing database objects are illegal.

BREAKING CHANGE:

The GraphQL validity type was relaxed to allow infinite (null)
from-dates. This is normally a non-breaking change, but the type's name
was also changed from RAValidityInput to RAOpenValidityInput. Since
validity on IT-systems was _just_ introduced, we presume that no-one is
relying on this yet, and choose to break it instead of introducing a new
GraphQL version.