feat(graphql): [#56602] add support for bitemporal facet writes in GraphQL
This commit adds support for bitemporal creates and updates, by making a
breaking change to the `facet_create` and `facet_update` endpoints and
their corresponding input models.

While this is in itself a breaking change, it is not marked as such, as
the following commit introduces a new GraphQL version to remedy this.
The commits are kept separate as they are logically distinct.