fix: [#57830] Remove new-org-unit-tag
`imported_org_unit_tag` is not used by anyone. If we would like to know
exacly which organizations are imported by this application, we should
put them in a specific root-org. This could be mapped in the json
dictionary.

We remove `imported_org_unit_tag` entirely because it makes the code
more robust. It is bad practise to consider two name strings equal when
there is an `IMPORTED FROM LDAP: ` string prepended to one of them.