feat(graphql): [#58028] introduce auditlog models filter enum
This commit changes the `auditlog`'s `models` filter type from an
unconstrained string to a structured enum.

The change is made as the sample space is already predetermined and
derived from the source code itself, not from the data.
Additionally having it as an enum allows for better IDE integration, as
GraphQL IDEs can show the possible choice values.