feat(graphql)!: [#56025] structured resolver filters
This commit introduces structured filters to GraphQL resolvers, such
that it is now possible to make a query alike:
```graphql
query FindPloneUser {
  itusers(
    filter: {
      itsystem: {user_keys: "Plone"},
      employee: {query: "Louise"}
    }
  ) {
    objects {
      current {
        user_key
        person {
          name
        }
      }
    }
  }
}
```
To find people named "Louise" that has IT users for the "Plone" ITSystem.

The result would be similar to:
```json
{
  "data": {
    "itusers": {
      "objects": [
        {
          "current": {
            "user_key": "LouiseL",
            "person": [
              {
                "name": "Louise Schæffer Larsen"
              }
            ]
          }
        },
        {
          "current": {
            "user_key": "LouiseC",
            "person": [
              {
                "name": "Louise Lundstrøm Christiansen"
              }
            ]
          }
        }
      ]
    }
  }
}
```

Before this change several queries had to be done client side, namely:

1. A query to find people UUIDs from the search query.
2. A query to find the UUID of the "Plone" itsystem.
3. A query to produce our result using the previously found UUIDs

```graphql
query FindPloneUser {
  itusers(
    filter: {
      employees: [
        # Result from first query here
        "cd64c199-a208-4a71-b334-6d896aa2f78d",
        "23bca421-9a36-4112-b3e4-70e53ddff96b"
      ]
      itsystem_uuids: [
        # Result from second query here
        "988dead8-7564-464a-8339-b7057bfa2665"
      ]
    }
  ) {
    objects {
      current {
        user_key
        person {
          name
        }
      }
    }
  }
}
```
Which all together would produce the same result as the new query.

At the same time all the old filters have been deprecated in favor of
the new ones, such that if you used `employees` before, you should now
rather use `employee.uuids` instead, such as:
```graphql
query FindPloneUser {
  itusers(
    filter: {
      employee: {
        uuids: [
          # Result from first query here
          "cd64c199-a208-4a71-b334-6d896aa2f78d",
          "23bca421-9a36-4112-b3e4-70e53ddff96b"
        ]
      }
      itsystem: {
        uuids: [
          # Result from second query here
          "988dead8-7564-464a-8339-b7057bfa2665"
        ]
      }
    }
  ) {
    objects {
      current {
        user_key
        person {
          name
        }
      }
    }
  }
}
```

The filtering options available under the `employee` key of `itusers`
top-level are the same that are available under the `filter` key of the
`employees` top-level.

BREAKING CHANGE:

It is no longer possible to filter on both UUIDs and user-keys in
GraphQL at the same time. This was never possible on top-level filters
only sub-filters, this odd behavior has been removed such that
sub-filters now behave the same way as top-level filters.