feat(test): [#xxxxx] created unit test for employees decorate
This was done in relation to another issue, but i noticed the test would not cover
the problem in that issue. But instead of removing the test I just moved it to this branch.
The test is still relevant, and we are missing more tests.. but we also have plans on
moving the responsibility of decorating search result data to the frontend, so more
tests on this autocomplete logic is properly not a good idea to use to much time on