feat(graphql): [#58392] implement model field and filter
This commit ensures that organisation functions are handled correctly
with regards to the registration model filter and field.

Previously an address registration with a given UUID would also be
returned as an engagement with the same ID as the code did not
distinguish the different organisation functions in the OIO standard.

The same applied to model filters, which would simply filter for any
organisation function, not for the specific organisation function
specified.

Now everything should be working correctly and the model field and
filter should reflect the overloaded nature of organisation functions
and their concrete models in OS2mo itself.