fix: [#57422] Create *and* edit IT users
Previously, any existing IT user with the same `user_key` would simply
be skipped, rather than updated. This made it impossible to add an
engagement UUID to an existing IT user.

This commit changes the return type of
'SyncTool.format_converted_objects` and thus breaks some tests. The
return type is now a list of `(MOBase, Verb)` tuples. `Verb` is a new
enum which can take the values `CREATE` and `EDIT` depending on which
operation is necessary to perform against the MO service API.