fix(graphql): [#58997] fix collections registrations
This commit fixes the `registrations` fields on the `Response` models.
These fields were unintentionally broken by OS2mo 32.12.0, when model
filters were reimplemented to handle organisation functions correctly.

The root-cause of the issue was a disabled test-case, as testing the
registrations subsystem is currently not possible due to the way that
database connections are configured during testing.

This commit simply fixes the issue, but does not protect against
regressions in the future, as testing is still troublesome, and handled
by another ticket, namely: 58757