feat(graphql): [#58997] allow querying at a set registration time
This commit introduces the ability to query data at a set registration
time, by providing the `registration_time` filter parameter.

To show how this feature is useful, consider the following timeline for
a fictitious person 'Birgitta':

* Birgitta is born in 1948 and bears the surname Duschek
* Birgitta gets married in 1978 and changes her surname to Munk
* Birgitta enters our system in 2019, but without her marriage record
* Birgitta notices the mistake in 2024 and gets it corrected

Using GraphQL we can query the registration history for Birgitta:
```graphql
query EmployeeRegistrations($uuid: UUID!) {
  employees(filter: {uuids: [$uuid]}) {
    objects {
      registrations {
        start
        end
      }
      uuid
    }
  }
}
```
Which yields:
```json
{
  "data": {
    "employees": {
      "objects": [
        {
          "registrations": [
            {
              "start": "2019-12-18T12:58:27.678744+00:00",
              "end": "2024-01-03T13:53:11.940575+00:00"
            },
            {
              "start": "2024-01-03T13:53:11.940575+00:00",
              "end": null
            }
          ],
          "uuid": "0004b952-a513-430b-b696-8d393d7eb2bb"
        }
      ]
    }
  }
}
```
We see two registrations, one for when Birgitta originally entered our
system and one for the correction to her surname.

We can also query her current data, using a query such as:
```graphql
query Employee($uuid: UUID!) {
  employees(
    filter: {uuids: [$uuid], from_date: null, to_date: null}
  ) {
    objects {
      objects {
        validity {
          from
          to
        }
        name
      }
    }
  }
}
```
Which yields:
```json
{
  "data": {
    "employees": {
      "objects": [
        {
          "objects": [
            {
              "validity": {
                "from": "1948-11-10T00:00:00+01:00",
                "to": "1978-01-02T00:00:00+01:00"
              },
              "name": "Birgitta Duschek"
            },
            {
              "validity": {
                "from": "1978-01-03T00:00:00+01:00",
                "to": null
              },
              "name": "Birgitta Munk"
            }
          ]
        }
      ]
    }
  }
}
```
Thus we can see Birgittas birthdate and the name she carried from birth
until her marriage in 1978.

Using this new feature we can however query our knowledge of Birgitta at
any point in the registration time, using a query such as:
```graphql
query EmployeeRegTime($uuid: UUID!, $registration_time: DateTime) {
  employees(
    filter: {
      uuids: [$uuid]
      registration_time: $registration_time,
      from_date: null, to_date: null,
    }
  ) {
    objects {
      objects {
        validity {
          from
          to
        }
        name
      }
    }
  }
}
```
Querying with `registration_time="2025-01-01"` yields the same as
before, as querying into the future is the same as querying now.

However querying with `registration_time="2020-01-01"` yields:
```json
{
  "data": {
    "employees": {
      "objects": [
        {
          "objects": [
            {
              "validity": {
                "from": "1948-11-10T00:00:00+01:00",
                "to": null
              },
              "name": "Birgitta Duschek"
            }
          ]
        }
      ]
    }
  }
}
```
Here we see Birgitta with her maiden name instead of her current
surname, although she got married 42 years prior.

This is because we had not yet gotten the surname correction into the
system by 2020. The correction only came in 2024, and as such, in 2020
we believed the above data to be correct.

And finally querying with `registration_time="2015-01-01"` yields:
```json
{
  "data": {
    "employees": {
      "objects": []
    }
  }
}
```
As Birgitta had not yet entered our system.

Thus this feature allows for querying data as-if we had done it on
another date in the past, thus allowing us to time travel.