feat(graphql): [#56025] allow clearing extension attributes
Before this change, running a mutation alike:
```graphql
mutation UpdateExtensionField(
  $uuid: UUID!,
  $start: DateTime!,
  $end: DateTime,
  $extension_3: String
) {
  engagement_update(
    input: {
      uuid: $uuid,
      validity: {from: $start, to: $end},
      extension_3: $extension_3
    }
  ) {
    uuid
  }
}
```
With arguments:
```json
{
  "uuid": "00e96933-91e4-42ac-9881-0fe1738b2e59",
  "start": "2000-01-01T00:00:00+00:00",
  "end": null,
  "extension_3": ""
}
```
Would not actually clear `extension_3` as the attributed would be
filtered by the `to_handler_dict` helper method.

This change ensures that only `None`s are filtered instead of all falsy
values. In the future all `UNSET` fields to be filtered, but that
requires a larger revamp to the GraphQL implementation moving away from
pydantic models and towards Strawberry models.