feat(auth): [#59181] support SIMPLE auth
This commit adds support for using SIMPLE auth which works by simply
sending the fully qualified DN of the client (user) and the clients
clear-text password (see RFC 2251 and RFC 2829) to the LDAP Server.

This mechanism has security problems because the password can be read
from the network traffic if a non-TLS secured connection is used.

However it is plenty fine to use for a local development setup or
integration tests, whereas the current authentification backend in NTLM
is Microsoft specific and thus will not work with standard LDAP
implementations.

In the future authentification should be done using SASL/GSSAPI.

To utilize SIMPLE auth, set the `ldap_auth_method` environmental
variable to `"SIMPLE"` as opposed to the default of `"NTLM"`.