fix: [#57700] GraphQL dates
Previously, the LoRa Connector was cached across the request without any
regard for the actual loraparams it was initialised with (because the
caching was applied to `_create_graphql_connector`, whose parameters
never change). Now, the LoRa Connector is cached *after* it has been
initialised with validity parameters. This means that the validity
parameter parsing logic is no longer cached, but at least it is correct.

Note that the only reason the LoRa Connector is cached in the first
place is to ensure the general DataLoader in lora.py works. When the
Service API is removed, and GraphQL refactored, the Connector caching
can be removed as well.