refactor(graphql): [#57700] refactor Response model argument
This commit removes the required `model` argument on the `Response`
generic model in favor of extracting the model from the generic
parameters.

This was previously stated as impossible as the reasoning behind the
introduction of the model field, but that was in fact incorrect as
proven by the existence of this change.