feat(config)!: [#59347] remove forbidden usernames file support
This commit changes the behavior of the:
`CONVERSION_MAPPING__USERNAME_GENERATOR__FORBIDDEN_USERNAMES`
environmental variable, which is used to specify outputs that the chosen
username generator cannot generate.

Prior to this commit the configuration variable could either contain
usernames to block, or filenames for files containing usernames to
block.

The distinction was made simply by whether the username in the
environmental variable list ended on `.csv` og `.txt`. This
functionality has now been removed, and thus usernames must be provided
directly in the environmental variable.

The goal of the change is to follow the 12-factor recommendation for
configuration and to avoid having to release new versions of the LDAP
integration simply to change customer configurations.

BREAKING CHANGE:

It used to be possible to mount configuration files to the mapping /
forbidden usernames folder and refer to them using the CONVERSION_MAP
environmental variable, however this is no longer possible.

If you need this behavior, simply convert your `.csv` or `.txt` file to
a json list and provide it in the above mentioned environmental variable.