fix(deps): [#58056] add missing cryptodome dependency
The dependency was removed in version 4.0.1 as it seemed unused by the
integration, as it turns it it was very much needed by the LDAP3
library, which does not declare its dependencies properly.

See: https://github.com/cannatag/ldap3/issues/1051

This commit readds the dependency and ensures that the integration will
function once again.