feat: [#59732] support single field mapping to primary engagement
This commit adds the `get_primary_engagement_dict` jinja function, which
returns the primary engagement for the provided employee UUID as a dict.

This allows changing only a single field on the primary engagement as
one can refer to to all the old / current field values via the returned
dictionary.

For example:
```yaml
ldap_to_mo:
  Engagement:
    objectClass: ramodels.mo.details.engagement.Engagement
    _import_to_mo_: "true"
    person: '{{ dict(uuid=employee_uuid or NONE) }}'

    # Change the user-key of the engagement to fit the sAMAccountName
    user_key: '{{ ldap.sAMAccountName }}'
    validity: '{{ dict(from_date=now()|mo_datestring) }}'

    # These fields remain unmodified and all follow the same form
    primary: '{{ dict(uuid=get_primary_engagement_dict(employee_uuid).primary_uuid) }}'
    org_unit: '...'
    job_function: '...'
    engagement_type: '...'
    uuid: '{{ get_primary_engagement_dict(employee_uuid).uuid }}'
```