feat: [#59792] support single field mapping on employees
This commit adds the `get_employee_dict` jinja function, which returns
the employee dictionary for the provided employee UUID as a dict.

This allows changing only a single field on the employee as one can refer
to to all the old / current field values via the returned dictionary.

For example:
```yaml
ldap_to_mo:
  Employee:
    objectClass: ramodels.mo.details.engagement.Employee
    _import_to_mo_: "true"
    cpr_no: "{{ ldap.employeeNumber|strip_non_digits or NONE }}"
    uuid: "{{ employee_uuid or NONE }}"

    # Change the user-key of the employee to fit the sAMAccountName
    user_key: "{{ ldap.sAMAccountName }}"

    # These fields remain unmodified
    givenname: "{{ get_employee_dict(employee_uuid).givenname }}"
    surname: "{{ get_employee_dict(employee_uuid).surname }}"
    nickname_givenname: "{{ get_employee_dict(employee_uuid).nickname_givenname }}"
    nickname_surname: "{{ get_employee_dict(employee_uuid).nickname_surname }}"
```