refactor: [#59804] use proper locking instead of sleeping
This commit refactors the `wait_for_import_to_finish` and
`wait_for_export_to_finish` functions to do proper locking instead of
just doing random sleeping while waiting to be the next in line to run.

While this saves a miniscule amount of CPU processing power, the main
goal of the change is assuring fairness of access to the shares
resource, such that a first-come, first-served principle is upheld
ensuring that events are handled in the proper order.