feat: [#59803] allow terminating details
This commit expands the entity mapping configuration to take a new
optional key: `_terminate_`, this key allows users to decide whether
(and when) a certain entity should be terminated.

If the key is set and is templated to a non-`None` value, the output
value should be a `mo_datestring` specifing the date at which the entity
should be terminated.

In order to ensure that terminatation template works, please ensure that
all ldap variables used in the termination is found in the reverse
mapping configuration, for instance by simply reverse mapping it to
`None` (Atleast until #59832 is resolved).

Termination only currently works for addresses, engagements and
it-users, but the code is flexible w.r.t. to supporting other ramodel
detail types in the future as required.