refactor(auth): [#59559] make the DIPEX client less special
This commit moves the DIPEX client from the ordinary realm-builder to
the integration realm builder, as DIPEX is nothing but a collection of
integrations and thus should not be treated differently than other
integrations.

This commit is the first step in eliminating the DIPEX client and
ensuring that each integration has its own client, thereby ensuring that
we can track which integration makes which change in MO, w.r.t. the
audit log.