feat!: [#59798] replace "roles" with role bindings
BREAKING CHANGE:

Before this change, we had "roles" which connected an employee with a
class that was connected to the "role_type" facet. This patch *removes*
roles from the legacy service API and from all GraphQL versions.
Empirically, "roles" were unused.

This change introduces "rolebindings" (to use a less ambiguous term) and
uses a new "role" facet instead of "role_type" so we can roll out the
change without destructing data. Role bindings differ from the old roles
in that they bind an IT User (instead of a person) with a role. Roles
are now supposed to be used as IAM, not assorted responsibilities. All
the deprecated fields on "roles" have been removed from "rolebindings"
and "org_unit" is now optional.