fix(filter): [#60126] fixup apply_discriminator input type
This commit fixes an issue with the `apply_discriminator` implementation,
prior to this commit the function wrongfully assumed the `search_result`
input to be `search_result["attributes"]` making the function entirely
inoperable.

This commit fixes us the defective tests to reflect the actual responses
returned by `ldap_connection.search` and fixes the code correspondingly.

The code is still defective however, as there are a plethora of codepaths
that use the `apply_discriminator` function without actually ensuring
that the `discriminator_field` is in the result from the search operation.
This issue will be resolved in a future commit.