fix: [#60147] fix JobTitleFromADToMO
This commit fixes the customer specific job-title extension within the
customer_specific module. This extension was unintentionally broken by
version 4.1.2 of the LDAP integration, where the integration was updated
to use GraphQL v21, while the extension was not.

The root-cause of the issue was many fold:

* The integration currently lacks integration tests

If we had, had integration tests, these would easily have detected the issue.

* The integration does not currently use the FastRAMQPI auto-generated client.

If we had, had the auto-generated client, bumping to a newer version
would have triggered a static check against the GraphQL schema, which
would have detected the issue.

This commit changes to code to using the auto-generated client to
mitigate this issue in the future.