test(ldap): [#60126] introduce ldap3 mock tests
It appears that a lot of the tests in the integration do not correctly
call the the underlying ldap library, leading to a lot of work reworking
tests or trying to replicate issues.

This commit introduces the first couple of tests that utilize the
mocking approach recommended my the ldap3 library, such that entities
are loaded into an in-memory ldap server emulation and replies are
generated by the library as if an actual ldap server was contacted.

At this time it is mainly a proof of concept for how to write proper
ldap tests in the future, and a huge amount of work still needs to be
put in to refactor all the existing tests.